Updated chatgpt to be in detailed report, updated api footer
