+++ src/Pages/DetailedQuestions.tsx
@@ -7,13 +7,12 @@ import { DetailedQuestion } from "../QuestionData/DetailedQuestion";
 import jsonData from "../QuestionData/DetailedQuestions.json";
 import "../Formatting/General.css";
 import "../Formatting/DetailedQuestions.css";
+import { keyData, saveKeyData, updateKeyData } from "../APIFooter";
 
 //local storage and API Key: key should be entered in by the user and will be stored in local storage (NOT session storage)
-let keyData = "";
-const saveKeyData = "MYKEY";
 const prevKey = localStorage.getItem(saveKeyData); //so it'll look like: MYKEY: <api_key_value here> in the local storage when you inspect
 if (prevKey !== null) {
-  keyData = JSON.parse(prevKey);
+  updateKeyData(JSON.parse(prevKey));
 }
 
 function DetailedQuestions() {
