+++ src/Pages/DetailedReport.tsx
@@ -6,8 +6,12 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import "../Formatting/Home.css";
 import "../Formatting/General.css";
 import "../Formatting/Report.css";
+import OpenAIAPi from "openai";
 import { keyData, saveKeyData, updateKeyData } from "../APIFooter";
-//import OpenAIRequest from './OpenAITest';
+
+let completion: OpenAIAPi.Chat.Completions.ChatCompletion;
+let openai = new OpenAIAPi({ apiKey: keyData, dangerouslyAllowBrowser: true });
+let responseData = "Placeholder!";
 
 //local storage and API Key: key should be entered in by the user and will be stored in local storage (NOT session storage)
 const prevKey = localStorage.getItem(saveKeyData); //so it'll look like: MYKEY: <api_key_value here> in the local storage when you inspect
@@ -28,6 +32,28 @@ function DetailedReport() {
     setKey(event.target.value);
   }
 
+  async function ChatGPT()
+  {
+    //Creates ChatGPT
+    openai = new OpenAIAPi({ apiKey: keyData, dangerouslyAllowBrowser: true });
+    //Queries ChatGPT
+    completion = await openai.chat.completions.create({
+      messages: [
+        { "role": "system", "content": "You are a helpful assistant designed to output JSON." },
+        { "role": "user", "content": "Who won the world series in 2020?" }
+      ],
+      model: "gpt-3.5-turbo",
+      response_format: { "type": "json_object" }
+    });
+    //Stores response data
+    if(completion.choices[0].message.content !== null){
+      responseData = completion.choices[0].message.content;
+    }
+    else{
+      responseData = "Error!";
+    }
+  }
+
   return (
     <div className={bodyClassName} id="bigBody">
       <header className="General-header">
@@ -43,15 +69,12 @@ function DetailedReport() {
       <div className="Page-body">
         <div className="Page-text">
           <p className="Report-header">View your Detailed Quiz Results!</p>
-          <p className="Report-body">
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
-            eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim
-            ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-            aliquip ex ea commodo consequat. Duis aute irure dolor in
-            reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-            pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-            culpa qui officia deserunt mollit anim id est laborum.
-          </p>
+          <Form className="Report-body">
+            <Button className="Button-ChatGPT" onClick={ChatGPT}>
+              Generate Report
+            </Button>
+          </Form>
+          {responseData}
         </div>
       </div>
